Add title to link back to exhibit page
Code now iterates thru the exhibit pages that contain the item,
and checks that $_SERVER['HTTP_REFERER'] points to one of these
pages. If so, the link will print, using as text the title of
the exhibit page. Also, since this is a "legacy" theme that
mimics the behavior of Omeka 1.5.3, where there were sections
with no landing page, we also compare with the first child