Modified code for link back to exhibit page in item page
Moved most of the orginal functionality added in a previous commit
from item.php into the CulCustomize plugin, since this
functionality is needed in all the themes.
Also, now the link to the exhibit page displays the page title
(including parent pages). Also, now the links to other pages
containing the item also display on top of the page, and the list
does not repeat the aforementioned link back to the exhibit page.