Change function call items_for_file to cul_items_for_file
The cul_items_for_file is a helper function, defined in
plugins/CulCustomize/CulCustomizePlugin.php. This helper
function contains a call to items_for_file with arguments
that are always the same, regardless of the theme. Therefore,
the call to items_for_file was moved to the cul_items_for_file
helper function. Any change to the arguments of the items_for_file
call now just need to be implemented inside the helper function,
instead of implemented in each theme.