Fix runoff text in Firefox
The descriptive text was running off the right hand side of
the page -- half the text was not visible. It seems that Firefox
does not handle width and max-width within elements in a table
the same way as Safari or Chrome. Change were made to the CSS to
give a width in terms of pixels to the content div within the table.
In addtion, some other changes were made to the layout.